requirements.txt
astroid==2.12.12
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
greenlet==2.0.1
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
numpy==1.23.4
pandas==1.5.1
platformdirs==2.5.3
pylint==2.15.5
pyodbc==4.0.34
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
